aaaaaa

+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: nait-bou <nait-bou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/04 10:00:12 by nait-bou          #+#    #+#             */
-/*   Updated: 2023/11/04 10:26:06 by nait-bou         ###   ########.fr       */
+/*   Updated: 2023/11/09 18:23:19 by nait-bou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (src[i] != '\0' && i < size -1)
